Add distinction between artists and clients in auth module

+++ src/app/Auth/DatabaseAuth.php
@@ -2,11 +2,14 @@
 
 namespace G1c\Culturia\app\Auth;
 
+use G1c\Culturia\app\Artists\model\ArtistsModel;
+use G1c\Culturia\app\Artists\table\ArtistsTable;
 use G1c\Culturia\app\Auth\model\ClientModel;
 use G1c\Culturia\app\Auth\table\ClientTable;
 use G1c\Culturia\framework\Auth;
 use G1c\Culturia\framework\Auth\User;
 use G1c\Culturia\framework\Database\NoRecordException;
+use G1c\Culturia\framework\Database\Table;
 use G1c\Culturia\framework\Session\SessionInterface;
 
 class DatabaseAuth implements Auth
@@ -15,12 +18,14 @@ class DatabaseAuth implements Auth
 
     private $user;
     private SessionInterface $session;
+    private ArtistsTable $artistsTable;
 
-    public function __construct(ClientTable $clientTable, SessionInterface $session)
+    public function __construct(ClientTable $clientTable, ArtistsTable $artistsTable,  SessionInterface $session)
     {
 
         $this->clientTable = $clientTable;
         $this->session = $session;
+        $this->artistsTable = $artistsTable;
     }
 
     public function getUser(): ?User
@@ -30,23 +35,30 @@ class DatabaseAuth implements Auth
         }
         $userId = $this->session->get('auth.user');
         if($userId) {
-            try {
-                $this->user = $this->clientTable->find($userId);
+            $this->user = $this->searchUser($this->clientTable, $userId);
+            if(!is_null($this->user)) {
                 return $this->user;
-            } catch (NoRecordException $e) {
-                $this->session->delete('auth.user');
-                return null;
             }
+            $this->user = $this->searchUser($this->artistsTable, $userId);
+            if(!is_null($this->user)) {
+                return $this->user;
+            }
+            $this->session->delete('auth.user');
         }
         return null;
     }
 
-    public function login(string $username, string $password): ?User
+    public function login(string $username, string $password, string $role): ?User
     {
         if(empty($username) || empty($password)) {
             return null;
         }
-        $user = $this->clientTable->makeQuery()
+        if($role) {
+            $table = $this->clientTable;
+        } else {
+            $table = $this->artistsTable;
+        }
+        $user  = $table->makeQuery()
             ->where("email = :email OR username = :email")
             ->params([':email' => $username])
             ->fetchOrFail();
@@ -60,9 +72,10 @@ class DatabaseAuth implements Auth
     public function register(array $params): ?User
     {
         if(!array_key_exists('email', $params)
-        ||!array_key_exists('password', $params)
+            ||!array_key_exists('password', $params)
             ||!array_key_exists('password2', $params)
             ||!array_key_exists('username', $params)
+            || !array_key_exists('role', $params)
 
         ) {
             return null;
@@ -76,10 +89,16 @@ class DatabaseAuth implements Auth
                 "modification_date" => date("Y-m-d H:i:s")
 
             ];
-            /** @var ClientModel $user */
-            $this->clientTable->insert($params);
-            $id = $this->clientTable->getPdo()->lastInsertId();
-            $user = $this->clientTable->find($id);
+            if($params["role"]) {
+                /** @var ClientModel $user */
+                $table =$this->clientTable;
+            } else {
+                /** @var ArtistsModel $user */
+                $table = $this->artistsTable;
+            }
+            $table->insert($params);
+            $id = $table->getPdo()->lastInsertId();
+            $user = $table->find($id);
             if($user){
                 $this->session->set('auth.user', $user->id);
                 return $user;
@@ -92,5 +111,15 @@ class DatabaseAuth implements Auth
     public function logout(): void
     {
         $this->session->delete('auth.user');
+    }
+
+    private function searchUser(Table $table, int $id): ?User
+    {
+        try {
+            return $table->find($id);
+        } catch (NoRecordException $e) {
+            return null;
+        }
+
     }
 }
